Fix check for PackageKit D-Bus specs
Currently, it checks if the `PK_INTERFACE_XML` or `PK_TRANSACTION_INTERFACE_XML`
variables are set, and not that the files with those paths exist.

If the `find_path` call fails, `PK_INTERFACE_XML` will be set to
`PK_INTERFACES_DIR-NOTFOUND/org.freedesktop.PackageKit.xml`, which isn't
empty (and likewise for `PK_TRANSACTION_INTERFACE_XML`), so the
condition is never true.